Small tweak to NamePtg and Ref3DPtg
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@696501 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFName.java
@@ -47,7 +47,7 @@ public final class HSSFName {
      * @return sheet name, which this named range referred to
      */
     public String getSheetName() {
-        short indexToExternSheet = _definedNameRec.getExternSheetNumber();
+        int indexToExternSheet = _definedNameRec.getExternSheetNumber();
 
         return _book.getWorkbook().findSheetNameFromExternSheet(indexToExternSheet);
     }
